refactor: moving the loganalytics/logic clients
