+++ azurerm/internal/services/loganalytics/client/client.go
@@ -1,4 +1,4 @@
-package loganalytics
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/preview/operationalinsights/mgmt/2015-11-01-preview/operationalinsights"
@@ -12,7 +12,7 @@ type Client struct {
 	WorkspacesClient     *operationalinsights.WorkspacesClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	WorkspacesClient := operationalinsights.NewWorkspacesClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&WorkspacesClient.Client, o.ResourceManagerAuthorizer)
