+++ azurerm/internal/services/logic/client/client.go
@@ -1,4 +1,4 @@
-package logic
+package client
 
 import (
 	"github.com/Azure/azure-sdk-for-go/services/logic/mgmt/2016-06-01/logic"
@@ -9,7 +9,7 @@ type Client struct {
 	WorkflowsClient *logic.WorkflowsClient
 }
 
-func BuildClient(o *common.ClientOptions) *Client {
+func NewClient(o *common.ClientOptions) *Client {
 	WorkflowsClient := logic.NewWorkflowsClientWithBaseURI(o.ResourceManagerEndpoint, o.SubscriptionId)
 	o.ConfigureClient(&WorkflowsClient.Client, o.ResourceManagerAuthorizer)
